# doc-files/kypo-user-and-group-swagger-open-api.yaml
...
swagger: "2.0"
info:
  version: "Version: 1.1.77"
  title: "KYPO User and Group - API Reference"
host: "localhost:8084"
basePath: "/kypo-rest-user-and-group/api/v1"
...
schemes:
- "http"
- "https"
paths:
  /groups:
    get:
      tags:
      - "groups"
      summary: "Get groups."
      description: ""
      operationId: "getGroups"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Filtering on IDMGroup entity attributes"
        required: false
        schema:
          $ref: "#/definitions/Predicate"
      - name: "fields"
        in: "query"
        description: "Fields which should be returned in REST API response"
        required: false
        type: "string"
      - name: "page"
        in: "query"
        description: "Results page you want to retrieve (0..N)"
        required: false
        type: "int"
        x-example: "0"
      - name: "size"
        in: "query"
        description: "Number of records per page."
        required: false
        type: "int"
        x-example: "20"
      - name: "sort"
        in: "query"
        description: "Sorting criteria in the format: property(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        type: "array"
        items:
          type: "string"
          example: "asc"
        collectionFormat: "multi"
        x-example: "asc"
      responses:
        200:
          description: "All groups found."
          schema:
            $ref: "#/definitions/GroupRestResource"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    post:
      tags:
      - "groups"
      summary: "Create new group."
      description: ""
      operationId: "createNewGroup"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Group to be created."
        required: true
        schema:
          $ref: "#/definitions/NewGroupDTO"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/GroupDTO"
        201:
          description: "Given group created."
          schema:
            $ref: "#/definitions/GroupDTO"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    put:
      tags:
      - "groups"
      summary: "Update group."
      description: ""
      operationId: "updateGroup"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Group to be updated."
        required: true
        schema:
          $ref: "#/definitions/UpdateGroupDTO"
      responses:
        200:
          description: "Group updated."
        405:
          description: "Group is external and cannot be modified."
          schema:
            $ref: "#/definitions/ApiError"
        409:
          description: "Name of the main group cannot be changed."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "groups"
      summary: "Delete groups"
      description: "Tries to deleteIDMGroup groups with given ids and returns groups\
        \ and statuses of their deletion"
      operationId: "deleteGroups"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Ids of groups to be deleted."
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
      responses:
        200:
          description: "Returned HTTP status OK."
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /groups/{groupId}:
    get:
      tags:
      - "groups"
      summary: "Get group with given id"
      description: ""
      operationId: "getGroupById"
      produces:
      - "application/json"
      parameters:
      - name: "groupId"
        in: "path"
        description: "Id of group to be returned."
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Group found."
          schema:
            $ref: "#/definitions/GroupDTO"
        404:
          description: "Group cannot be found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "groups"
      summary: "Delete group"
      description: "Tries to deleteIDMGroup group with given id and returns if it\
        \ was successful."
      operationId: "deleteGroup"
      produces:
      - "application/json"
      parameters:
      - name: "groupId"
        in: "path"
        description: "Id of group to be deleted."
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Returned deletion status of the group."
        404:
          description: "Group cannot be found."
          schema:
            $ref: "#/definitions/ApiError"
        405:
          description: "Group cannot be deleted because it is a main group."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
  /groups/{id}/roles:
    get:
      tags:
      - "groups"
      summary: "Returns all roles of group"
      description: ""
      operationId: "getRolesOfGroup"
      parameters:
      - in: "body"
        name: "body"
        description: "Filtering on IDMGroup entity attributes"
        required: false
        schema:
          $ref: "#/definitions/Predicate"
      - name: "id"
        in: "path"
        description: "id"
        required: true
        type: "integer"
        format: "int64"
      - name: "page"
        in: "query"
        description: "Results page you want to retrieve (0..N)"
        required: false
        type: "int"
        x-example: "0"
      - name: "size"
        in: "query"
        description: "Number of records per page."
        required: false
        type: "int"
        x-example: "20"
      - name: "sort"
        in: "query"
        description: "Sorting criteria in the format: property(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        type: "array"
        items:
          type: "string"
          example: "asc"
        collectionFormat: "multi"
        x-example: "asc"
      responses:
        200:
          description: "All roles of group found."
          schema:
            $ref: "#/definitions/RoleRestResource"
        404:
          description: "Group cannot be found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
  /users:
    get:
      tags:
      - "users"
      summary: "Gets all users."
      description: ""
      operationId: "getUsers"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Filtering on User entity attributes"
        required: false
        schema:
          $ref: "#/definitions/Predicate"
      - name: "fields"
        in: "query"
        description: "Fields which should be returned in REST API response"
        required: false
        type: "string"
      - name: "page"
        in: "query"
        description: "Results page you want to retrieve (0..N)"
        required: false
        type: "int"
        x-example: "0"
      - name: "size"
        in: "query"
        description: "Number of records per page."
        required: false
        type: "int"
        x-example: "20"
      - name: "sort"
        in: "query"
        description: "Sorting criteria in the format: property(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        type: "array"
        items:
          type: "string"
          example: "asc"
        collectionFormat: "multi"
        x-example: "asc"
      responses:
        200:
          description: "Users found."
          schema:
            $ref: "#/definitions/UserRestResource"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
    delete:
      tags:
      - "users"
      summary: "DeleteUsers"
      description: "Delete users based on given ids."
      operationId: "deleteUsers"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Ids of users to be deleted."
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
      responses:
        200:
          description: "Returned HTTP status OK."
        404:
          description: "User could not be found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /users/groups:
    get:
      tags:
      - "users"
      summary: "Gets users in given groups."
      description: ""
      operationId: "getUsersInGroups"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Filtering on User entity attributes"
        required: false
        schema:
          $ref: "#/definitions/Predicate"
      - name: "fields"
        in: "query"
        description: "Fields which should be returned in REST API response"
        required: false
        type: "string"
      - name: "ids"
        in: "query"
        description: "Ids of groups where users are assigned."
        required: true
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "page"
        in: "query"
        description: "Results page you want to retrieve (0..N)"
        required: false
        type: "int"
        x-example: "0"
      - name: "size"
        in: "query"
        description: "Number of records per page."
        required: false
        type: "int"
        x-example: "20"
      - name: "sort"
        in: "query"
        description: "Sorting criteria in the format: property(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        type: "array"
        items:
          type: "string"
          example: "asc"
        collectionFormat: "multi"
        x-example: "asc"
      responses:
        200:
          description: "Users in specific groups found."
          schema:
            $ref: "#/definitions/UserRestResource"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /users/ids:
    get:
      tags:
      - "users"
      summary: "Gets users with given ids."
      description: "Page size cannot be higher than 1000"
      operationId: "getUsersWithGivenIds"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Filtering on User entity attributes"
        required: false
        schema:
          $ref: "#/definitions/Predicate"
      - name: "fields"
        in: "query"
        description: "Fields which should be returned in REST API response"
        required: false
        type: "string"
      - name: "ids"
        in: "query"
        description: "Ids of users to be obtained."
        required: true
        type: "array"
        items:
          type: "integer"
          format: "int64"
        collectionFormat: "multi"
      - name: "page"
        in: "query"
        description: "Results page you want to retrieve (0..N)"
        required: false
        type: "int"
        x-example: "0"
      - name: "size"
        in: "query"
        description: "Number of records per page."
        required: false
        type: "int"
        x-example: "20"
      - name: "sort"
        in: "query"
        description: "Sorting criteria in the format: property(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        type: "array"
        items:
          type: "string"
          example: "asc"
        collectionFormat: "multi"
        x-example: "asc"
      responses:
        200:
          description: "Users with given ids found."
          schema:
            $ref: "#/definitions/UserRestResource"
        404:
          description: "User could not be found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /users/info:
    get:
      tags:
      - "users"
      summary: "Get user info"
      description: "Returns details of user who is logged in."
      operationId: "getUserInfo"
      parameters: []
      responses:
        200:
          description: "Base user info found."
          schema:
            $ref: "#/definitions/UserDTO"
        404:
          description: "User could not be found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
  /users/not-in-groups/{groupId}:
    get:
      tags:
      - "users"
      summary: "Gets all users except users in given group."
      description: ""
      operationId: "getAllUsersNotInGivenGroup"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Filtering on User entity attributes"
        required: false
        schema:
          $ref: "#/definitions/Predicate"
      - name: "groupId"
        in: "path"
        description: "Id of group whose users do not getGroupById."
        required: true
        type: "integer"
        format: "int64"
      - name: "fields"
        in: "query"
        description: "Fields which should be returned in REST API response"
        required: false
        type: "string"
      - name: "page"
        in: "query"
        description: "Results page you want to retrieve (0..N)"
        required: false
        type: "int"
        x-example: "0"
      - name: "size"
        in: "query"
        description: "Number of records per page."
        required: false
        type: "int"
        x-example: "20"
      - name: "sort"
        in: "query"
        description: "Sorting criteria in the format: property(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        type: "array"
        items:
          type: "string"
          example: "asc"
        collectionFormat: "multi"
        x-example: "asc"
      responses:
        200:
          description: "User found."
          schema:
            $ref: "#/definitions/UserRestResource"
        404:
          description: "Some user could not be found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
      security:
      - bearerAuth: []
...
securityDefinitions:
  bearerAuth:
    description: "Obtain bearer token from OIDC provider and enter the token in the\
      \ following field. Use format 'Bearer {token}'"
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  GroupRestResource:
    type: "object"
    properties:
      content:
        type: "array"
        description: "Retrieved IDMGroups from databases."
        items:
          $ref: "#/definitions/GroupDTO"
      pagination:
        description: "Pagination including: page number, number of elements in page,\
          \ size, total elements and total pages."
        $ref: "#/definitions/Pagination"
    description: "Content (Retrieved data) and meta information about REST API result\
      \ page. Including page number, number of elements in page, size of elements,\
      \ total number of elements and total number of pages."
  NewGroupDTO:
    type: "object"
    required:
    - "description"
    - "name"
    properties:
      name:
        type: "string"
        example: "Main group"
        description: "A name of the group."
      description:
        type: "string"
        example: "Group for main users."
        description: "A description of the group."
      expirationDate:
        type: "string"
        format: "date-time"
        example: "2019-11-20T10:28:02.727Z"
        description: "Time until the group is valid."
      users:
        type: "array"
        description: "List of users who is assigned to group."
        uniqueItems: true
        items:
          $ref: "#/definitions/UserForGroupsDTO"
      groupIdsOfImportedUsers:
        type: "array"
        example: "[1]"
        description: "Main identifiers of group."
        items:
          type: "integer"
          format: "int64"
    description: "Data that need to be provided to create a new group."
...
  Predicate:
    type: "object"
...